move _private to cli.usage
